Pixiv: refactored (most of the) fetch code
Previously the fetching of items from Pixiv has been heavily sequence-based: the retrieval of infinitely many objects as needed was encoded, and then manually limited to a certain number per page.

This is an issue however, because it significantly increases the search's load times. It is better to return lots of small pages, so Mihon can immediately display them and estimate how many more to fetch.

Aside from that, we're now caching more and better. There's still room for improvement tho. The code for different fetch components is also grouped together, which should improve readability.